Warnings (/W4) as errors and fixed warnings

+++ RaytracingInAWeekend/camera.h
@@ -10,7 +10,7 @@ class camera
 public:
 	camera(glm::vec3 lookfrom, glm::vec3 lookat, glm::vec3 upVec, float vfov, float aspect_ratio)
 	{
-		theta = vfov * M_PI/ 180.f;
+		theta = vfov * static_cast<float>(M_PI)/ 180.f;
 		half_height = tan(theta / 2.f);
 		half_width = aspect_ratio * half_height;
 		look_at = lookat;
@@ -36,8 +36,8 @@ public:
 		glm::vec3 u = normalize(cross(up, w));
 		glm::vec3 v = cross(w, u);
 		lower_left_corner = origin - half_width * u - half_height * v - w;
-
-		horizontal = 2 * half_width * u;
+	
+		horizontal = 2.f * half_width * u;
 		vertical = 2.f * half_height * v;
 	}
